backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
motor>=2.5.1
pydantic>=1.8.2
python-dotenv>=0.19.0
scikit-learn>=1.0.0
openai>=1.0.0